requirements.txt
BeautifulSoup==3.2.1
Fabric==1.4.3
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
boto==2.5.2
git-remote-helpers==0.1.0
latimes-calculate==0.1.8
pusher==0.6
pycrypto==2.6
python-dateutil==1.5
python-elections==0.35
ssh==1.7.14
wsgiref==0.1.2
